requirements.txt
cryptography==42.0.2
bcrypt==4.1.2
pandas==2.2.1
openpyxl==3.1.2
PySide6==6.6.1
python-dotenv==1.0.1
pyinstaller==6.5.0 